refactor: rename folders in src
