+++ src/components/burger-ingredients/burger-ingredients-section/burger-ingredient/burger-ingredient.module.scss
@@ -1,4 +1,4 @@
-@import '../../index.scss';
+@import '../../../../index.scss';
 
 .card {
   position: relative;
