+++ src/components/burger-ingredients/burger-ingredients-section/burger-ingredients-section.module.scss
@@ -1,4 +1,4 @@
-@import '../../index.scss';
+@import '../../../index.scss';
 
 .heading {
   @include show-font(
