+++ src/components/ingredient-details/ingredient-details.jsx
@@ -2,7 +2,7 @@ import { useSelector } from 'react-redux';
 
 import getCurrentIngredient from '../../services/features/current-ingredient/selectors';
 
-import styles from './IngredientDetails.module.scss';
+import styles from './ingredient-details.module.scss';
 
 function IngredientDetails() {
   const currentIngredient = useSelector(getCurrentIngredient);
